Fixed import errors related to moving do_checks. Created an additional helper class and moved two classes outside Misc.py
